Add diagonal operations for mpi exponential Pauli Z gate
Note that N-qubit exponential Pauli Z gate with N>=3 does not support
diagonal operation.
